+++ backend/src/hatchling/utils/constants.py
@@ -1,1 +1,2 @@
 DEFAULT_BUILD_SCRIPT = 'hatch_build.py'
+DEFAULT_CONFIG_FILE = 'hatch.toml'